requirements.txt
aws-msk-iam-sasl-signer-python==1.0.1
confluent-kafka==2.4.0